+++ src/main/java/racingcar/Car.java
@@ -20,6 +20,4 @@ public class Car {
     public int getPosition() {
         return position;
     }
-
-
 }
