+++ src/main/java/racingcar/Cars.java
@@ -13,6 +13,14 @@ public class Cars {
         }
         return carList;
     }
-
-
+    public static List<String> createCarList(String[] carNames){
+        List<String> carNameList = new ArrayList<>();
+        for(String carName: carNames){
+            if(carNameList.contains(carName)){
+                throw new IllegalArgumentException("중복된 이름입니다.");
+            }
+            carNameList.add(carName);
+        }
+        return carNameList;
+    }
 }
