+++ src/main/java/racingcar/CheckValidation.java
@@ -1,0 +1,18 @@
+package racingcar;
+
+public class CheckValidation {
+    public static void validateName(String[] carNames){
+        for(String carName:carNames){
+            if(carName.length() > 5){
+                throw new IllegalArgumentException("5자리 이하로 입력해주세요.");
+            }
+        }
+    }
+    public static void validateNum(String roundNum){
+        try {
+            Integer.parseInt(roundNum);
+        }catch (Exception e){
+            throw new IllegalArgumentException("숫자를 입력하세요.");
+        }
+    }
+}
