+++ src/main/java/racingcar/RoundNumberInput.java
@@ -6,15 +6,8 @@ public class RoundNumberInput {
     public static Integer roundNumInput(){
         System.out.println("시도할 회수는 몇회인가요?");
         String stringRoundNum = Console.readLine();
-        validateNum(stringRoundNum);
+        CheckValidation.validateNum(stringRoundNum);
         return Integer.parseInt(stringRoundNum);
     }
 
-    public static void validateNum(String roundNum){
-        try {
-            Integer.parseInt(roundNum);
-        }catch (Exception e){
-            throw new IllegalArgumentException("숫자를 입력하세요.");
-        }
-    }
 }
